npmのアップデート

+++ src/components/mypage/MobileMypage.tsx
@@ -1,11 +1,15 @@
 import { Link } from "@chakra-ui/next-js";
 import { Card, CardBody, HStack, LinkProps, Spacer, Stack, StackDivider, Text } from "@chakra-ui/react";
-import { ReactNode } from "react";
+import { Loading } from "components/common/Loading";
 import { ChangeCreaterModeButton } from "components/mypage/ChangeCreaterModeButton";
+import { useCreaterModeContext } from "components/provider/CreaterModeProvider";
+import { ReactNode } from "react";
 
 type Props = {};
 
 export const MobileMypage = ({}: Props) => {
+  const { createrMode } = useCreaterModeContext();
+
   type CardLinkProps = {
     children: ReactNode;
     href: string;
@@ -21,23 +25,26 @@ export const MobileMypage = ({}: Props) => {
       </Link>
     );
   };
-
-  return (
-    <>
-      <HStack spacing="20" align="start">
-        <Stack width="100%">
-          <Card bgColor="black">
-            <CardBody>
-              <Stack spacing={4} divider={<StackDivider />}>
-                <CardLink href="changeEmail">メールアドレスの確認･変更</CardLink>
-                <CardLink href="changePassword">パスワードの変更</CardLink>
-                <CardLink href="deleteAccount">退会</CardLink>
-              </Stack>
-            </CardBody>
-          </Card>
-          <ChangeCreaterModeButton />
-        </Stack>
-      </HStack>
-    </>
-  );
+  if (createrMode === null) {
+    return <Loading />;
+  } else {
+    return (
+      <>
+        <HStack spacing="20" align="start">
+          <Stack width="100%">
+            <Card bgColor="black">
+              <CardBody>
+                <Stack spacing={4} divider={<StackDivider />}>
+                  <CardLink href="changeEmail">メールアドレスの確認･変更</CardLink>
+                  <CardLink href="changePassword">パスワードの変更</CardLink>
+                  <CardLink href="deleteAccount">退会</CardLink>
+                </Stack>
+              </CardBody>
+            </Card>
+            {createrMode ? <></> : <ChangeCreaterModeButton />}
+          </Stack>
+        </HStack>
+      </>
+    );
+  }
 };
